Feedback on content prior to the growth matrix
I've changed _ex:_ to _e.g. (exempli gratia)_ throughout (it's [recommended English usage](https://english.stackexchange.com/questions/16197/whats-the-difference-between-e-g-and-ex) and [much more common today](https://books.google.com/ngrams/graph?content=ex%2Ce.g.&year_start=1800&year_end=2000&corpus=15&smoothing=3&share=&direct_url=t1%3B%2Cex%3B%2Cc0%3B.t1%3B%2Ce.g.%3B%2Cc0)).

Additional feedback:

> Be proficient in at least one **core** programming language (ex: JavaScript, Ruby);

"Core" is unclear, even with the brief list that follows.  Do you mean core to the industry, or languages used extensively in our product?  If the list is very short (JavaScript and Ruby) should we just say "Be proficient in JavaScript or Ruby?"  Would we hire a very senior engineer that had no experience with JavaScript or Ruby? If so, maybe "Gain proficiency" instead of "Be proficient?"

> Have strong computer science fundamentals (ex: object-oriented programming, data structures, algorithms)

This still needs more definition, as we've discussed a bit internally.

> * Write correct, well-tested, high-quality code to deliver end to end features;
> * Deploy and support features in production (we deploy every day);

Three concerns in these two statements:

1. "to deliver features" feels like it misses the underlying motivations: To deliver value, to solve problems, to support our mission.
2. "to deliver end to end features" and "Deploy and support features in production" feel somewhat redundant.
3. The editorializing "(we deploy every day)" feels unnecessary.

I'd rewrite these two statements to ensure each has a clear objective.  Something like:

* Write correct, well-tested, high-quality code to support our mission; (focus on quality and value of code produced)
* Own technical design, development, deployment, and support of features; (focus on end-to-end responsibility)

> * Demonstrate passion for the Code.org mission.
> * Demonstrate support of our mission of enabling “all students in all schools to learn computer science”.

Are these redundant? Or is the second one a subtle call-out to "disagree and commit" as a key practice that we want to see in day-to-day work?  Maybe we should be more explicit about that, especially if we want this document to feel more generally applicable outside of our organization.

> It is also part of our culture for peers and development managers to encourage and facilitate sharing of growth-oriented and constructive feedback.

I can't parse this sentence. Is the "it" at the beginning referring to the growth matrix, or does it refer to the "part of our culture" described in the rest of the statement?  If the former, I'd say something like "It also facilitates sharing of growth-oriented and constructive feedback from peers and development managers."  If the latter, maybe "We also encourage peers and development managers to share growth-oriented and constructive feedback, and facilitate this twice a year."